RIP to Float16, it's sad that Intel doesn't support you.

+++ Sources/Voxelotl/Chunk.swift
@@ -123,7 +123,7 @@ public struct Chunk: Hashable {
 
 public enum BlockType: Hashable {
   case air
-  case solid(_ color: Color<Float16>)
+  case solid(_ color: Color<Float>)
 }
 
 public struct Block: Hashable {
